cinematch/pyproject.toml
[project]
name = "cinematch"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "endee>=0.1.18",
    "google-genai>=1.66.0",
    "google-generativeai>=0.8.6",
    "pandas>=2.3.3",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sentence-transformers>=5.2.3",
    "streamlit>=1.55.0",
    "tenacity>=9.1.4",
    "tqdm>=4.67.3",
]
